+++ src/Endpoint/ServicesInterface.php
@@ -8,6 +8,7 @@ use Aceproxies\ResellerApi\Exception\ApiException;
 use Aceproxies\ResellerApi\Exception\InvalidResponseException;
 use Aceproxies\ResellerApi\Exception\TransportException;
 use Aceproxies\ResellerApi\Request\Service\UpdateServiceRequest;
+use Aceproxies\ResellerApi\Response\Service\BandwidthResponse;
 use Aceproxies\ResellerApi\Response\Service\DetailResponse;
 use Aceproxies\ResellerApi\Response\Service\ListResponse;
 use InvalidArgumentException;
@@ -29,6 +30,14 @@ interface ServicesInterface
      * @throws InvalidArgumentException
      */
     public function find(string $code): ?DetailResponse;
+
+    /**
+     * @throws ApiException
+     * @throws InvalidResponseException
+     * @throws TransportException
+     * @throws InvalidArgumentException
+     */
+    public function getBandwidth(string $serviceCode): ?BandwidthResponse;
 
     /**
      * @throws ApiException